add excitation tests to structure
